Fixed review comments
Added 16 characters restriction on passwords
Renamed ed25519Key to publicKeyMaterialKey
Renamed ValidEd25519PrivKey to ValidateEd25519PrivKey
Fixed protected len check in UnprotectSymKey

+++ scripts/unittest.sh
@@ -1,7 +1,15 @@
 #!/bin/sh
 
 # -race increase test time a lot with crypto things, so the timeout must take that into account
-go test -v -timeout 60s -race ./... -coverprofile cover.out.tmp
+go test -timeout 60s -race ./... -coverprofile cover.out.tmp
+
+# Stop here if test have failed, as coverage below will shift
+# the test failures up and make it easy to miss.
+if [ $? -ne 0 ]; then
+    echo "FAIL - Some test have failed."
+    exit 1
+fi
+
 cat cover.out.tmp | grep -v "types.go" > cover.out
 rm -f cover.out.tmp
 # coverage report
